refactor: make `ComboboxArea` component reusable
